+++ src/Fetchers/FIHFetcher/FIHFetcher.ts
@@ -1,8 +1,6 @@
 import { Fetcher, FetcherOptions } from "../Fetcher.js";
 import { Competition } from "../../Objects/Competition.js";
 import { Match } from "../../Objects/Match.js";
-import { ICSCreator } from "../../Utils/ICSCreator.js";
-import { Gender } from "../../Objects/Gender.js";
 import { Official } from "../../Objects/Official.js";
 import { FIHCompetitionFetcher } from "./FIHCompetitionFetcher.js";
 import { FIHMatchFetcher } from "./FIHMatchFetcher.js";
@@ -40,45 +38,6 @@ export class FIHFetcher extends Fetcher {
 
         this.competitionFetcher = new FIHCompetitionFetcher(this);
         this.matchFetcher = new FIHMatchFetcher(this);
-    }
-
-    /**
-     * Fetch the matches from FIH.
-     */
-      protected async fetch() {
-        this.log("info", "Fetching competitions.");
-        const competitions = await this.fetchCompetitions();
-        const promises = [];
-
-        this.log("info", `Found ${competitions.size} competitions.`);
-        this.log("info", "Fetching matches and creating competition files.");
-
-        for (const competition of competitions.values()) {
-            // Fetch match for every competition
-            const matchPromise = this.fetchMatches(competition);
-            matchPromise.then(result => {
-                competition.getMatches().push(...result.values());
-                return ICSCreator.createCompetitionICS(competition);
-            });
-
-            promises.push(matchPromise);
-        }
-
-        // Wait for all matches to fetch
-        await Promise.all(promises);
-        const competitionsArray = Array.from(competitions.values());
-
-        // Create total calendar files.
-        await Promise.all([
-            ICSCreator.createTotalICS(this, competitionsArray),
-            ICSCreator.createGenderTotalICS(this, competitionsArray,
-                Gender.MEN),
-            ICSCreator.createGenderTotalICS(this, competitionsArray,
-                Gender.WOMEN)
-        ]);
-
-        this.log("info", "Finished.");
-        return competitionsArray;
     }
 
     /**
