+++ src/Utils/ICSCreator.ts
@@ -46,6 +46,8 @@ export class ICSCreator {
         let matches = competitions.map(e => e.getMatches()).flat();
         matches = matches.filter(m => m.getGender() === gender);
 
+        if (matches.length === 0) return;
+
         const path = `${this.genderToString(gender, true)}-matches`;
         const title = `All ${fetcher.getAbbr()} ${this.genderToString(
             gender, false)} matches`;
